requirements.txt
openai
imapclient
python-dotenv
beautifulsoup4
